backend/requirements.txt
# FastAPI & Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=14.0
python-multipart==0.0.9

# Database
supabase>=2.11.0
asyncpg>=0.30.0
pgvector==0.2.4
SQLAlchemy==2.0.36
redis>=5.0.0

# AI / OpenAI
openai>=1.58.1,<2.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
tiktoken>=0.7.0

# Document Parsing
pymupdf>=1.24.0
pdfplumber>=0.11.0
docx2txt>=0.8

# AWS
boto3==1.34.34
botocore==1.34.34
aiosmtplib==3.0.1

# Auth & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Config
pydantic-settings==2.2.1
email-validator>=2.0.0
python-dotenv==1.0.1

# HTTP client
httpx>=0.26.0,<0.29.0

# Data Processing
numpy>=1.26.2

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
